schedule_views.pyへ関数を移動
